pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "cgbv"
version = "0.1.0"
description = "Cross-Granularity Boundary Verification for NL-to-Logic Formalization Faithfulness"
requires-python = ">=3.11"
dependencies = [
    "z3-solver>=4.13",
    "openai>=1.30",
    "httpx>=0.27",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "python-dotenv>=1.0",
    "rich>=13.7",
]

[project.scripts]
cgbv = "cgbv.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["cgbv*"]
